Details Update
Setup a way to return the model to a variable to make it easier to work directly with the methods available.

A new model for product options help it easier to grab just the options for an item instead of all options for entire transaction, but you can still do both. check the docs.